20/10/25

+++ src/Svc.Core/Cor.HRMM/Models/DTOs/JgStepDto.cs
@@ -4,14 +4,14 @@ public class JgStepListDto : BaseDto
 {
     public Guid JobGradeId { get; set; } = default!; // JobGrade
     public string Name { get; set; } = default!;
-    public decimal Salary { get; set; } = default!;
+    public double Salary { get; set; } = default!;
     public string JobGrade { get; set; } = default!;
 }
 
 public class JgStepAddDto
 {
     public string Name { get; set; } = default!;
-    public decimal Salary { get; set; } = default!;
+    public double Salary { get; set; } = default!;
     public Guid JobGradeId { get; set; } = default!; // JobGrade
 }
 
@@ -19,7 +19,7 @@ public class JgStepModDto
 {
     public Guid Id { get; set; }
     public string Name { get; set; } = default!;
-    public decimal Salary { get; set; } = default!;
+    public double Salary { get; set; } = default!;
     public Guid JobGradeId { get; set; } = default!; // JobGrade
     public string RowVersion { get; set; } = default!;
 }
