+++ src/Svc.Core/Cor.HRMM/Models/DTOs/PositionReqDto.cs
@@ -7,7 +7,7 @@ public class PositionReqListDto : BaseDto
     public string Gender { get; set; } = default!; //enum.PositionGender (0/1)
     public string SaturdayWorkOption { get; set; } = default!; //enum.WorkOption (0/1)
     public string SundayWorkOption { get; set; } = default!; //enum.WorkOption (0/1)
-    public decimal WorkingHours { get; set; } = default!;
+    public double WorkingHours { get; set; } = default!;
     public string GenderStr { get; set; } = default!;
     public string SaturdayWorkOptionStr { get; set; } = default!;
     public string SundayWorkOptionStr { get; set; } = default!;
@@ -21,7 +21,7 @@ public class PositionReqAddDto
     public string Gender { get; set; } = default!; //enum.PositionGender (0/1)
     public string SaturdayWorkOption { get; set; } = default!; //enum.WorkOption (0/1)
     public string SundayWorkOption { get; set; } = default!; //enum.WorkOption (0/1)
-    public decimal WorkingHours { get; set; } = default!;
+    public double WorkingHours { get; set; } = default!;
     public Guid ProfessionTypeId { get; set; } = default!; //Lup.ProfessionType
     public Guid PositionId { get; set; } = default!; // Position
 }
@@ -32,7 +32,7 @@ public class PositionReqModDto
     public string Gender { get; set; } = default!; //enum.PositionGender (0/1)
     public string SaturdayWorkOption { get; set; } = default!; //enum.WorkOption (0/1)
     public string SundayWorkOption { get; set; } = default!; //enum.WorkOption (0/1)
-    public decimal WorkingHours { get; set; } = default!;
+    public double WorkingHours { get; set; } = default!;
     public Guid ProfessionTypeId { get; set; } = default!; //Lup.ProfessionType
     public Guid PositionId { get; set; } = default!; // Position
     public string RowVersion { get; set; } = default!;
