+++ src/Svc.Core/Cor.HRMM/Models/Entities/JgStep.cs
@@ -3,7 +3,7 @@
 public class JgStep: BaseEntity
 {
     public string Name { get; set; } = default!;
-    public decimal Salary { get; set; } = default!;
+    public double Salary { get; set; } = default!;
     public Guid JobGradeId { get; set; } = default!;
 
     //******************************************//
