+++ src/Svc.Core/Cor.HRMM/Models/Entities/JobGrade.cs
@@ -3,6 +3,6 @@
 public class JobGrade: BaseEntity
 {
     public string Name { get; set; } = default!;
-    public decimal StartSalary { get; set; } = default!;
-    public decimal MaxSalary { get; set; } = default!;
+    public double StartSalary { get; set; } = default!;
+    public double MaxSalary { get; set; } = default!;
 }
