+++ src/Svc.Core/Cor.HRMM/Models/Entities/PositionReq.cs
@@ -5,7 +5,7 @@ public class PositionReq : BaseEntity
     public string Gender { get; set; } = default!; //PositionGender
     public string SaturdayWorkOption { get; set; } = default!; //WorkOption
     public string SundayWorkOption { get; set; } = default!; //WorkOption
-    public decimal WorkingHours { get; set; } = default!;
+    public double WorkingHours { get; set; } = default!;
     public Guid ProfessionTypeId { get; set; } = default!; //Lup.ProfessionType
     public Guid PositionId { get; set; } = default!; // Position
